delete Set property of node

+++ src/main/java/cn/icedsoul/cutter/algorithm/CutGraphAlgorithm.java
@@ -7,5 +7,5 @@ import java.util.Map;
 
 public interface CutGraphAlgorithm {
 
-    Map<Integer, List<Long>> calculate();
+    Map<Integer, List<Integer>> calculate();
 }
